+++ components/Board/Board.tsx
@@ -1,19 +1,19 @@
 import Link from 'next/link'
+// import PriorityContainer from './PriorityContainer/PriorityContainer'
 import CommonContainer from './CommonContainer/CommonContainer'
-import PriorityContainer from './PriorityContainer/PriorityContainer'
 
 export default function Board() {
     return (
         <section className='w-screen p-8'>
             <div className="flex justify-between">
-				<h1 className="text-4xl font-semibold">Public Board</h1>
-				<Link
-					href="/note/new"
-					className="bg-[#994BFF] rounded-sm flex items-center text-sm font-semibold px-3 border-[#994BFF] border hover:text-[#994BFF] hover:bg-white">
-					Create new Note
-				</Link>
-			</div>
-            <PriorityContainer />
+                <h1 className="text-4xl font-semibold">Public Board</h1>
+                <Link
+                    href="/note/new"
+                    className="bg-[#994BFF] rounded-sm flex items-center text-sm font-semibold px-3 border-[#994BFF] border hover:text-[#994BFF] hover:bg-white">
+                    Create new Note
+                </Link>
+            </div>
+            {/* <PriorityContainer /> */}
             <CommonContainer />
         </section>
     )
