+++ components/Navbar/Navbar.tsx
@@ -3,7 +3,7 @@ import Link from "next/link";
 
 export default function Navbar() {
     return (
-        <section className="flex flex-col items-center h-screen p-4 border-r-2 border-gray-800">
+        <section className="flex flex-col items-center h-screen w-3/12 p-4 border-r-2 border-gray-800">
             <div className="flex items-center">
                 <Image src="/logo.png" width={50} height={50} alt="turtle-todo" className="" />
                 <h3 className="text-2xl font-semibold ml-2 text-[#994BFF]">Turtle Todos</h3>
